requirements.txt
pyyaml
memory-profiler==0.61.0
psutil==5.9.4